# optimize `cfg_simplify!`