--- a/test/compiler/irpasses.jl
+++ b/test/compiler/irpasses.jl
@@ -1568,3 +1568,5 @@ let m = Meta.@lower 1 + 1
 
     Core.Compiler.verify_ir(ir)
 end
+
+# JET.test_opt(Core.Compiler.cfg_simplify!, (Core.Compiler.IRCode,))