> Before

```julia
julia> @benchmark CC.cfg_simplify!(ir) setup=(ir = CC.copy(Main.ir))
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  196.333 μs …  22.691 ms  ┊ GC (min … max): 0.00% … 98.80%
 Time  (median):     208.792 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   219.824 μs ± 390.126 μs  ┊ GC (mean ± σ):  4.18% ±  2.40%

               ▄█▃
  ▁▁▁▁▁▁▁▁▁▂▂▃▆███▄▄▅▇▇▅▄▄▄▅▄▄▄▄▃▃▃▃▃▂▂▂▂▂▂▂▂▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  196 μs           Histogram: frequency by time          235 μs <

 Memory estimate: 169.48 KiB, allocs estimate: 4005.
```

> After

```julia
julia> @benchmark CC.cfg_simplify!(ir) setup=(ir = CC.copy(Main.ir))
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  49.541 μs …  31.882 ms  ┊ GC (min … max):  0.00% … 99.66%
 Time  (median):     54.042 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   61.492 μs ± 392.747 μs  ┊ GC (mean ± σ):  10.36% ±  1.72%

              ▄█▇▄▁
  ▁▁▁▁▁▂▂▃▄▄▅██████▇▅▅▄▃▃▃▃▄▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  49.5 μs         Histogram: frequency by time         67.2 μs <

 Memory estimate: 72.19 KiB, allocs estimate: 808.
```

with

```julia
julia> ir, = only(Base.code_ircode(sin, (Float64,));
```