@nanosoldier 2023-10-31
[Your benchmark job](https://github.com/JuliaLang/julia/pull/51958#issuecomment-1787966390) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/f2fdb64_vs_f631597/report.md).