@aviatesk 2023-11-01
The `cfg_simplify!` isn't being used in the current compiler implementation actually, so the benchmark results remain unchanged. #51960 will be the first one to integrate it into the compiler.